fix(a5): pgTAP -- cast explicito de bigint/smallint em create_opportunity
Mesmo problema de resolucao de assinatura: 5000000/300000/60 como
literais integer (nao bigint/smallint) ao lado de outros parametros ja
tipados nao bateram com a assinatura real da funcao sem cast explicito.

+++ supabase/tests/database/10_a5_pipeline.test.sql
@@ -337,7 +337,7 @@ select throws_ok(
 
 set local role authenticated;
 select set_config('request.jwt.claims', json_build_object('sub', :'bruno', 'role', 'authenticated')::text, true);
-select create_opportunity(:'lead_dois'::uuid, null::uuid, null::uuid, 5000000, 'fixed'::public.fee_model, 60) as opp_dois \gset
+select create_opportunity(:'lead_dois'::uuid, null::uuid, null::uuid, 5000000::bigint, 'fixed'::public.fee_model, 60::smallint) as opp_dois \gset
 select (get_opportunity(:'opp_dois'::uuid) ? 'value_cents') as owner_ve_exato \gset
 select ok(:'owner_ve_exato'::boolean, 'owner vê a chave value_cents (exata)');
 
@@ -371,7 +371,7 @@ select is(
 -- viewer: nenhuma chave financeira, em nenhum caminho.
 set local role authenticated;
 select set_config('request.jwt.claims', json_build_object('sub', :'ana', 'role', 'authenticated')::text, true);
-select create_opportunity(:'lead_sem_resp'::uuid, null::uuid, null::uuid, 300000, 'fixed'::public.fee_model) as opp_viewer_alvo \gset
+select create_opportunity(:'lead_sem_resp'::uuid, null::uuid, null::uuid, 300000::bigint, 'fixed'::public.fee_model) as opp_viewer_alvo \gset
 
 set local role authenticated;
 select set_config('request.jwt.claims', json_build_object('sub', :'elisa', 'role', 'authenticated')::text, true);
